:safety_vest: 후기 작성자만 수정/삭제 가능하도록 수정
- 서비스 레이어 예외처리 추가
  - 일반 사용자는 본인이 작성한 후기만 수정/삭제 가능하도록 권한 체크 예외처리 추가
- 문화행사별 후기 목록 조회 응답 DTO 변경
  - 클라이언트 단에서 후기별로 수정/삭제 가능 여부를 판단할 수 있도록 isEditable 불린 추가

+++ src/main/java/org/example/what_seoul/service/board/BoardService.java
@@ -10,15 +10,18 @@ import org.example.what_seoul.controller.board.dto.*;
 import org.example.what_seoul.controller.user.dto.ReqUpdateUserInfoDTO;
 import org.example.what_seoul.domain.board.Board;
 import org.example.what_seoul.domain.citydata.event.CultureEvent;
+import org.example.what_seoul.domain.user.RoleType;
 import org.example.what_seoul.domain.user.User;
 import org.example.what_seoul.exception.CustomValidationException;
 import org.example.what_seoul.repository.board.BoardRepository;
 import org.example.what_seoul.repository.citydata.event.CultureEventRepository;
 import org.example.what_seoul.service.user.UserService;
+import org.example.what_seoul.service.user.dto.LoginUserInfoDTO;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -58,7 +61,8 @@ public class BoardService {
         Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Direction.DESC, "createdAt"));
         Slice<Board> boardSlice = boardRepository.findAllByCultureEventId(cultureEventId, pageable);
 
-        Slice<ResGetBoardDTO> result = boardSlice.map(ResGetBoardDTO::from);
+        LoginUserInfoDTO loginUserInfo = userService.getLoginUserInfo();
+        Slice<ResGetBoardDTO> result = boardSlice.map(board -> ResGetBoardDTO.from(board, loginUserInfo));
 
         return new CommonResponse<>(true, "장소별 문화행사 후기 목록 조회 성공", result);
     }
@@ -66,6 +70,11 @@ public class BoardService {
     @Transactional
     public CommonResponse<ResUpdateBoardDTO> updateBoard(Long id, ReqUpdateBoardDTO req) {
         Board board = boardRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("문화행사 후기를 찾을 수 없습니다. 후기 id = " + id));
+
+        User user = (User) userService.getAuthenticationPrincipal();
+        if (user.getRole() == RoleType.USER && !Objects.equals(board.getUser().getId(), user.getId())) {
+            throw new AccessDeniedException("수정 권한이 없습니다.");
+        }
 
         // 기존 후기 내용
         String currContent = board.getContent();
@@ -102,6 +111,11 @@ public class BoardService {
     @Transactional
     public CommonResponse<ResDeleteBoardDTO> deleteBoard(Long id) {
         Board board = boardRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("문화행사 후기를 찾을 수 없습니다. 후기 id = " + id));
+
+        User user = (User) userService.getAuthenticationPrincipal();
+        if (user.getRole() == RoleType.USER && !Objects.equals(board.getUser().getId(), user.getId())) {
+            throw new AccessDeniedException("삭제 권한이 없습니다.");
+        }
         ResDeleteBoardDTO resDTO = ResDeleteBoardDTO.from(board); // LAZY 로딩 방지
 
         boardRepository.delete(board);
